강의영상 시청재게 기능 추가

+++ 온라인 강의 플랫폼/lecture/src/main/java/online/lecture/entity/MemberLecture.java
@@ -26,13 +26,13 @@ public class MemberLecture {
     private double progressRate;
 
     private Long LastWatchedVideoId;
-    private int LastWatchedVideoTime;
+    private double LastWatchedVideoTime;
 
     public void setLastWatchedVideoId(Long lastWatchedVideoId) {
         LastWatchedVideoId = lastWatchedVideoId;
     }
 
-    public void setLastWatchedVideoTime(int lastWatchedVideoTime) {
+    public void setLastWatchedVideoTime(double lastWatchedVideoTime) {
         LastWatchedVideoTime = lastWatchedVideoTime;
     }
